Add QueryFuture.Execute/Stream, delete RPCExecutor
