+++ rpc/QueryFuture.java
@@ -19,13 +19,100 @@
 
 package grakn.client.rpc;
 
+import grakn.client.common.exception.GraknClientException;
+import grakn.protocol.TransactionProto;
+
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
+import java.util.function.Function;
+import java.util.stream.StreamSupport;
 
-public interface QueryFuture<T> extends Future<T> {
+public abstract class QueryFuture<T> implements Future<T> {
     @Override
-    T get();
+    public abstract T get();
 
     @Override
-    T get(long timeout, TimeUnit unit);
+    public abstract T get(long timeout, TimeUnit unit);
+
+    static class Execute<T> extends QueryFuture<T> {
+        private final RPCResponseAccumulator collector;
+        private final Function<TransactionProto.Transaction.Res, T> responseReader;
+
+        public Execute(final RPCResponseAccumulator collector, final Function<TransactionProto.Transaction.Res, T> responseReader) {
+            this.collector = collector;
+            this.responseReader = responseReader;
+        }
+
+        @Override
+        public boolean cancel(final boolean mayInterruptIfRunning) {
+            return false; // Can't cancel
+        }
+
+        @Override
+        public boolean isCancelled() {
+            return false; // Can't cancel
+        }
+
+        @Override
+        public boolean isDone() {
+            return collector.isDone();
+        }
+
+        @Override
+        public T get() {
+            try {
+                final TransactionProto.Transaction.Res res = collector.take();
+                return responseReader.apply(res);
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+                throw new GraknClientException(e);
+            }
+        }
+
+        @Override
+        public T get(final long timeout, final TimeUnit unit) {
+            try {
+                final TransactionProto.Transaction.Res res = collector.take(timeout, unit);
+                return responseReader.apply(res);
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+                throw new GraknClientException(e);
+            }
+        }
+    }
+
+    static class Stream<T> extends QueryFuture<java.util.stream.Stream<T>> {
+        private final RPCIterator<T> iterator;
+
+        public Stream(final RPCIterator<T> iterator) {
+            this.iterator = iterator;
+        }
+
+        @Override
+        public boolean cancel(boolean mayInterruptIfRunning) {
+            return false; // Can't cancel
+        }
+
+        @Override
+        public boolean isCancelled() {
+            return false; // Can't cancel
+        }
+
+        @Override
+        public boolean isDone() {
+            return iterator.isStarted();
+        }
+
+        @Override
+        public java.util.stream.Stream<T> get() {
+            iterator.startIterating();
+            return StreamSupport.stream(((Iterable<T>) () -> iterator).spliterator(), false);
+        }
+
+        @Override
+        public java.util.stream.Stream<T> get(long timeout, TimeUnit unit) {
+            iterator.startIterating(timeout, unit);
+            return StreamSupport.stream(((Iterable<T>) () -> iterator).spliterator(), false);
+        }
+    }
 }
